+++ api/v2alpha1/nifiregistryclient_conversion.go
@@ -1,0 +1,4 @@
+package v2alpha1
+
+// Hub marks this type as the conversion hub.
+func (*NifiRegistryClient) Hub() {}